Fix a bug when you had multiple presets with the same name
You can have presets with the same name (but different types), such as
a configure preset and a build preset with the same name. The
positions of the multiple presets with the same name were being
computed as the position of the first occurrence. This is now fixed.

Also, we now prepend the preset type in the imenu entry name, since
imenu does not seem to like multiple entries with the same
name (consult-imenu worked fine, but not regular imenu).